refactor: Component iles naming & structure
