+++ tools/server/webui/src/lib/components/app/dialogs/DialogChatError.svelte
