+++ tools/server/webui/src/lib/components/app/dialogs/DialogConfirmation.svelte
